refactor: prefer wifi.txt at root of the device
Placing the wifi.txt at the root of the device is more user-friendly - you don't need to hunt around for where to place the file, and it allows using the same sd card on other devices. We still support the old, pak-specific file, but migrate it for users so they don't have to.